Actulizando Proyecto: creando nueva cabezera de Mascota.h y Mascota.cpp
