+++ Mascota.h
@@ -1,0 +1,7 @@
+#ifndef MASCOTA_H
+#define MASCOTA_H
+
+void RegistrarMascotas();
+void ListarMascotas();
+
+#endif
