+++ Sistema de Gestion Veterinaria.dev
@@ -29,7 +29,7 @@ IncludeVersionInfo=0
 SupportXPThemes=0
 CompilerSet=0
 CompilerSettings=0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;0;20;0;0;0
-UnitCount=9
+UnitCount=11
 
 [VersionInfo]
 Major=1
@@ -132,6 +132,26 @@ BuildCmd=
 
 [Unit9]
 FileName=propietario.cpp
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit10]
+FileName=Mascota.h
+CompileCpp=1
+Folder=
+Compile=1
+Link=1
+Priority=1000
+OverrideBuildCmd=0
+BuildCmd=
+
+[Unit11]
+FileName=Mascota.cpp
 CompileCpp=1
 Folder=
 Compile=1
