+++ menu.cpp
@@ -3,7 +3,7 @@
 #include "variablesGlobales.h"
 #include <iostream>
 #include "propietario.h"
-
+#include "Mascota.h"
 using namespace std;
 
 void Menu() {
@@ -165,6 +165,7 @@ void menuMascotas() {
             	RegistrarMascotas();
                 break;
             case 2:
+            	ListarMascotas();
                 break;
             case 3:
 				break;    
@@ -180,27 +181,6 @@ void menuMascotas() {
     } while (opcion != 5);
 }
 
-//Se añadió la funcion void registrar Mascotas
-void RegistrarMascotas() {
 
-    if (cantidadDeMascotas >= capacidadMaxima) {
-        cout << "No se pueden agregar más productos, límite alcanzado.\n";
-        
-        return;
-    }
-
-    Mascota m; //para la id
-    m.id_mascota = cantidadDeMascotas + 1;
-    cin.ignore();
-    cout << "\nIngrese nombre de la mascota: "; getline(cin, m.nombre);
-    cout<<"Ingrese la edad de la mascota: "; cin>> m.edad;
-    cout << "Ingrese la especie: "; cin >> m.especie;
-    cout<<"Ingrese la raza: "; cin>>m.raza;
-    mascota[cantidadDeMascotas] = m;
-    cantidadDeMascotas++;
-    cout << "\nMascota agregado exitosamente.\n\n";
-    
-    system("pause");
-}
